+++ {{cookiecutter.project_name}}/frontend/app/_draft.js
@@ -1,3 +1,0 @@
-import CatchAllPage from '../page';
-
-export default CatchAllPage;